Update 0.9 Обновлен метод. Проведен рефакторинг

+++ Program.cs
@@ -10,9 +10,9 @@ namespace IntroductionCSharp
          Console.WriteLine("Программа определения максимального числа из двух целых чисел");
          Console.WriteLine("-------------------------------------------------------------");
          Console.WriteLine("Введите первое число: ");
-         double a = Convert.ToDouble(Console.ReadLine());
+         int a = Convert.ToInt32(Console.ReadLine());
          Console.WriteLine("Введите второе число: ");
-         double b = Convert.ToDouble(Console.ReadLine());
+         int b = Convert.ToInt32(Console.ReadLine());
          Console.WriteLine("Первое число = " + a + ";");
          Console.WriteLine("Второе число = " + b + ";");
          if (a == b)
